Fix t/apache/subprocess.t on Windows, using httpd-2.4.x.
This test fails every time for me, but not with httpd-2.2.x, which I don't understand. Debugging, if you start up the server, then add a break point in APR's read_with_timeout() (in readwrite.c) and walk through the function each time it is hit then the test passes. Without debugging it puts a "APR::PerlIO::read: (70007) The timeout specified has expired" error in the error_log and fails all tests.

Presumably it is trying to read from the process spawned in subprocess.pm before the pipe is ready to be read from, and the delay in manually debugging fixes that. The comments in read_data() in subprocess.pm note that a delay is indeed required, and is achieved by select() in the non-PerlIO case (which I imagine nobody uses now anyway) or a apr_wait_for_io_or_timeout() call from apr_file_read() in the PerlIO case.

However, whilst that remark is true of the *nix implementation of apr_file_read() (in file_io/unix/readwrite.c), it is not true of the Win32 version (in file_io/win32/readwrite.c).

Thus, adding a short sleep() fixes it on Win32. What I don't understand is why this wasn't necessary with httpd-2.2.x. The file_io/win32/readwrite.c file is identical in the two versions I'm using (httpd-2.4.6 vs httpd-2.2.25; apr-1.4.8 in both).

git-svn-id: https://svn.apache.org/repos/asf/perl/modperl/branches/httpd24threading@1565275 13f79535-47bb-0310-9956-ffa450edef68